@@ -1,6 +1,4 @@
-'use strict'
-
-const findIn = (root, ...tags) => {
+export const findIn = (root, ...tags) => {
 	if (!root) return null
 	let el = root
 	for (let tag of tags) {
@@ -12,12 +10,12 @@ const findIn = (root, ...tags) => {
 	return el
 }
 
-const textOf = (el) => {
+export const textOf = (el) => {
 	if (!el || !el.children) return null
 	const c = el.children.find(c => c.type === 'text')
 	return c && c.value || null
 }
 
-const attrOf = (el, attr) => el && el.attributes && el.attributes[attr] || null
-
-module.exports = {findIn, textOf, attrOf}
+export const attrOf = (el, attr) => {
+	return el && el.attributes && el.attributes[attr] || null
+}
