@@ -1,12 +1,13 @@
-'use strict'
+import peek from 'peek-stream'
+import iconv from 'iconv-lite'
+import {stringify} from 'qs'
+import Promise from 'pinkie-promise'
+import _createFetch from 'fetch-ponyfill'
+const {fetch} = _createFetch({Promise})
+import createDebug from 'debug'
+import xmlParser from 'xml-reader'
 
-const peek = require('peek-stream')
-const iconv = require('iconv-lite')
-const {stringify} = require('qs')
-const Promise = require('pinkie-promise')
-const {fetch} = require('fetch-ponyfill')({Promise})
-const debug = require('debug')('query-fis-broker-wfs')
-const xmlParser = require('xml-reader')
+const debug = createDebug('query-fis-broker-wfs')
 
 const createDecoder = () => {
 	return peek({maxBuffer: 100}, (data, cb) => {
@@ -20,11 +21,11 @@ const createDecoder = () => {
 const isObject = o => 'object' === typeof o && !Array.isArray(o)
 const hasProp = (o, k) => Object.prototype.hasOwnProperty.call(o, k)
 
-const defaults = {
+export const defaults = {
 	userAgent: 'https://github.com/derhuerst/query-fis-broker-wfs'
 }
 
-const query = (endpoint, query, parsers, opt = {}) => {
+export const request = (endpoint, query, parsers, opt = {}) => {
 	if ('string' !== typeof endpoint || !endpoint) {
 		throw new Error('endpoint must be a non-empty string.')
 	}
@@ -96,5 +97,3 @@ const query = (endpoint, query, parsers, opt = {}) => {
 		})
 	})
 }
-
-module.exports = query
